refactor kodu
