+++ server/CMakeLists.txt
@@ -5,7 +5,7 @@ set(CMAKE_CXX_STANDARD 14)
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x")
 
-add_executable(server protbuf/messages.pb.cc main.cpp main.h messages.h messages.cpp)
+add_executable(server protbuf/messages.pb.cc main.cpp main.h messages.h messages.cpp Client.cpp Client.h)
 
 target_link_libraries(server -pthread -I/usr/local/include -L/usr/local/lib -lprotobuf -pthread -lpthread -lcrypto)
