+++ server/messages.h
@@ -10,28 +10,11 @@ const uint8_t HASH_SIZE[StorageCloud::HashAlgorithm_ARRAYSIZE] = {
         MD5_DIGEST_LENGTH,
 };
 
-using namespace std;
-using namespace StorageCloud;
-
-class Client {
-private:
-    int socket;
-    EncryptionAlgorithm encryption_algorithm = EncryptionAlgorithm::NOENCRYPTION;
-    string username = "";
-    HashAlgorithm hashing_algorithm = HashAlgorithm::H_SHA512;
-
-    bool getNBytes(int, uint8_t*);
-    uint32_t parseSize(const uint8_t*);
-    bool processMessage(uint8_t*, int);
-    bool parseMessage(uint8_t*, int, uint8_t*, uint8_t**, uint32_t*);
-    bool processCommand(Command*);
-    bool processHandshake(Handshake*);
-    bool sendServerResponse(const ServerResponse*);
-    bool prepareDataToSend(uint8_t*, uint32_t);
-
-public:
-    Client(int);
-    void loop(volatile bool*);
-};
+void calculateHash(uint8_t, const uint8_t*, int, uint8_t**, uint16_t*);
+bool compareHash(const uint8_t*, uint16_t, const uint8_t*, uint16_t);
+uint32_t parseSize(const uint8_t*);
+std::string printHash(const uint8_t, const uint8_t*);
+void encrypt(StorageCloud::EncryptionAlgorithm, const uint8_t*, uint32_t, uint8_t**, uint32_t*);
+void decrypt(StorageCloud::EncryptionAlgorithm, const uint8_t*, uint32_t, uint8_t**, uint32_t*);
 
 #endif //SERVER_MESSAGES_H
